+++ Code/class10.java
@@ -1,14 +1,14 @@
-public class class17{
+package Code;
+public class class10{
     public static void main(String[] args){
-        int x=1;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (i>=1 && i<=5 && j>=1 && j<=5){
-                System.out.print(" 0 ");
-                x=x-1;
-            } else {
+            if (j==11 && i==11){
                 System.out.print(" . ");
-            }
+                
+            } else {
+                    System.out.print(" 0 ");
+                }
         }
          System.out.println();
     }
