+++ Code/class11.java
@@ -1,8 +1,9 @@
-public class class10{
+package Code;
+public class class11{
     public static void main(String[] args){
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==11 && i==11){
+            if (j==11 || i==11){
                 System.out.print(" . ");
                 
             } else {
