+++ Code/class13.java
@@ -1,9 +1,12 @@
-public class class4{
+package Code;
+public class class13{
     public static void main(String[] args){
+        int x=0;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (i==11 && j==11){
+            if (j==2||i==2||j==20||i==20){
                 System.out.print(" 0 ");
+                x=x-1;
             } else {
                 System.out.print(" . ");
             }
