+++ Code/class14.java
@@ -1,9 +1,12 @@
-public class class2{
+package Code;
+public class class14{
     public static void main(String[] args){
+        int x=0;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (i==20){
+            if (j>=2 && j<=5 || i>=2 && i<=5){
                 System.out.print(" 0 ");
+                x=x-1;
             } else {
                 System.out.print(" . ");
             }
