+++ Code/class17.java
@@ -1,9 +1,10 @@
-public class class12{
+package Code;
+public class class17{
     public static void main(String[] args){
-        int x=0;
+        int x=1;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==i+1 ||j==i-1){
+            if (i>=1 && i<=5 && j>=1 && j<=5){
                 System.out.print(" 0 ");
                 x=x-1;
             } else {
