+++ Code/class19.java
@@ -1,14 +1,21 @@
-public class class3{
+package Code;
+public class class19{
     public static void main(String[] args){
+        int x=5;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j>=8 && j<=14){
+            
+            if (j==x-1){
                 System.out.print(" 0 ");
+                x= x+4;
+                
             } else {
                 System.out.print(" . ");
+                
             }
         }
          System.out.println();
+         x=x-21;
     }
 
 }
