+++ Code/class3.java
@@ -1,8 +1,9 @@
-public class class1{
+package Code;
+public class class3{
     public static void main(String[] args){
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==2){
+            if (j>=8 && j<=14){
                 System.out.print(" 0 ");
             } else {
                 System.out.print(" . ");
