+++ Code/class4.java
@@ -1,11 +1,10 @@
-public class class7{
+package Code;
+public class class4{
     public static void main(String[] args){
-        int x=21;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==x){
+            if (i==11 && j==11){
                 System.out.print(" 0 ");
-                x=x-1;
             } else {
                 System.out.print(" . ");
             }
