+++ Code/class5.java
@@ -1,3 +1,4 @@
+package Code;
 public class class5{
     public static void main(String[] args){
     for (int i=1; i<=21;i= i+1){
