+++ Code/class6.java
@@ -1,20 +1,15 @@
-public class class18{
+package Code;
+public class class6{
     public static void main(String[] args){
-        int x=5;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            
-            if (j==x-1){
+            if (j==i){
                 System.out.print(" 0 ");
-                x= x+4;
-                
             } else {
                 System.out.print(" . ");
-                
             }
         }
          System.out.println();
-         x=x-21;
     }
 
 }
