+++ Code/class7.java
@@ -1,14 +1,15 @@
-public class class8{
+package Code;
+public class class7{
     public static void main(String[] args){
-        int x=22;
+        int x=21;
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==i || j==x-i){
+            if (j==x){
                 System.out.print(" 0 ");
-                
+                x=x-1;
             } else {
-                    System.out.print(" . ");
-                }
+                System.out.print(" . ");
+            }
         }
          System.out.println();
     }
