+++ Code/class9.java
@@ -1,12 +1,14 @@
-public class class6{
+package Code;
+public class class9{
     public static void main(String[] args){
     for (int i=1; i<=21;i= i+1){
         for (int j=1;j<=21;j=j+1){
-            if (j==i){
+            if (j==j){
                 System.out.print(" 0 ");
+                
             } else {
-                System.out.print(" . ");
-            }
+                    System.out.print(" . ");
+                }
         }
          System.out.println();
     }
